serverless/openimage-auth-lambda/requirements.txt
boto3==1.24.61
botocore==1.27.61
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.1
google-auth==2.11.0
idna==3.3
jmespath==1.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.2
requests==2.28.1
rsa==4.9
s3transfer==0.6.0
six==1.16.0
urllib3==1.26.12
